app/requirements.txt
python-dotenv==1.0.0
flask==2.0.2
Flask-Cors==3.0.10

grpcio==1.50.0
grpcio-tools==1.50.0

werkzeug==2.0.3
pyjwt